+++ troposphere/haze/include/haze/usb_session.hpp
@@ -41,7 +41,7 @@ namespace haze {
 
             bool GetConfigured() const;
             Event *GetCompletionEvent(UsbSessionEndpoint ep) const;
-            Result TransferAsync(UsbSessionEndpoint ep, void *buffer, size_t size, u32 *out_urb_id);
+            Result TransferAsync(UsbSessionEndpoint ep, void *buffer, u32 size, u32 *out_urb_id);
             Result GetTransferResult(UsbSessionEndpoint ep, u32 urb_id, u32 *out_transferred_size);
             Result SetZeroLengthTermination(bool enable);
     };
